Switched back to Double. What I had forgotten is that the output buffer from AVAudio is indeed expecting Floats, so conversion is necessary. Gemini knew about vDSP.convertElements so that's the fix. I also switched to the 64-bit versions of random generation in Noise.

+++ ProgressionPlayer/Sources/Tones/Arrow.swift
@@ -8,7 +8,7 @@
 import Accelerate
 import AVFAudio
 
-typealias CoreFloat = Float
+typealias CoreFloat = Double
 
 class Arrow11 {
   // these are arrows with which we can compose (arr/arrs run first, then this arrow)
